add procfile and local_server script
I like to have a script in every project called local_server that
will boot the application, not matter what stack it is on or which
services need to be initialized so that there is no ambiguity about
launching a local instance. We should also use Procfile.dev by default
since it is almost always necessary to run multiple processes and for
those who prefer running them in the same shell, this makes it trivial.